+++ tests/datavalue_test.py
@@ -6,7 +6,8 @@ from pytest import mark, raises
 from wikidata.client import Client
 from wikidata.commonsmedia import File
 from wikidata.datavalue import DatavalueError, Decoder
-from wikidata.entity import Entity
+from wikidata.entity import Entity, EntityId
+from wikidata.globecoordinate import GlobeCoordinate
 from wikidata.multilingual import MonolingualText
 
 
@@ -86,6 +87,7 @@ def test_decoder__time(datatype: str, fx_client: Client):
         })
     assert (datetime.date(2017, 2, 22) ==
             d(fx_client, datatype, other_value(precision=11)))
+    assert 2017 == d(fx_client, datatype, other_value(precision=9))
     utc = datetime.timezone.utc
     assert (datetime.datetime(2017, 2, 22, 2, 53, 12, tzinfo=utc) ==
             d(fx_client, datatype, valid))
@@ -127,7 +129,7 @@ def test_decoder__time(datatype: str, fx_client: Client):
         d(fx_client, datatype, other_value(precision=None))
         # precision field is missing
     for p in range(1, 15):
-        if p in (11, 14):
+        if p in (9, 11, 14):
             continue
         with raises(DatavalueError):
             d(fx_client, datatype, other_value(precision=p))
@@ -151,3 +153,21 @@ def test_decoder_commonsMedia__string(fx_client: Client):
           {'value': 'The Fabs.JPG', 'type': 'string'})
     assert isinstance(f, File)
     assert f.title == 'File:The Fabs.JPG'
+
+
+def test_decoder_globecoordinate(fx_client: Client):
+    d = Decoder()
+    decoded = d(fx_client, 'globe-coordinate', {
+        'value': {
+            "latitude": 70.1525,
+            "longitude": 70.1525,
+            "precision": 0.0002777777777777778,
+            "globe": "http://www.wikidata.org/entity/Q111"
+        },
+        'type': 'globecoordinate'
+    })
+    gold = GlobeCoordinate(70.1525,
+                           70.1525,
+                           fx_client.get(EntityId("Q111")),
+                           0.0002777777777777778,)
+    assert decoded == gold