+++ wikidata/datavalue.py
@@ -16,10 +16,11 @@ but only need to satify::
 """
 import collections.abc
 import datetime
-from typing import TYPE_CHECKING, Mapping, Union
+from typing import TYPE_CHECKING, Any, Mapping, Union
 
 from .client import Client
 from .commonsmedia import File
+from .globecoordinate import GlobeCoordinate
 from .multilingual import MonolingualText
 if TYPE_CHECKING:
     from .entity import Entity  # noqa: F401
@@ -149,7 +150,8 @@ class Decoder:
     def time(self,
              client: Client,
              datavalue: Mapping[str, object]) -> Union[datetime.date,
-                                                       datetime.datetime]:
+                                                       datetime.datetime,
+                                                       int]:
         value = datavalue['value']
         if not isinstance(value, collections.abc.Mapping):
             raise DatavalueError(
@@ -195,6 +197,9 @@ class Decoder:
             precision = value['precision']
         except KeyError:
             raise DatavalueError('precision field is missing', datavalue)
+        if precision == 9:
+            # The time only specifies the year.
+            return int(time[1:5])
         if precision == 11:
             return datetime.date(int(time[1:5]), int(time[6:8]),
                                  int(time[9:11]))
@@ -205,7 +210,7 @@ class Decoder:
             ).replace(tzinfo=datetime.timezone.utc)
         else:
             raise DatavalueError(
-                '{!r}: time precision other than 11 or 14 is '
+                '{!r}: time precision other than 9, 11 or 14 is '
                 'unsupported'.format(precision),
                 datavalue
             )
@@ -215,6 +220,22 @@ class Decoder:
                         datavalue: Mapping[str, object]) -> MonolingualText:
         pair = datavalue['value']
         return MonolingualText(pair['text'], pair['language'])  # type: ignore
+
+    def globecoordinate(self,
+                        client: Client,
+                        datavalue: Mapping[str, Any]) -> GlobeCoordinate:
+        pair = datavalue['value']
+        # Try to split out the entity from the globe string.
+        globe_entity = pair["globe"].split("http://www.wikidata.org/entity/")
+        if len(globe_entity) != 2:
+            raise DatavalueError(
+                "Globe string {} does not appear to be a "
+                "valid WikiData entity URL".format(pair["globe"]))
+        entity_id = globe_entity[1]
+        return GlobeCoordinate(pair['latitude'],
+                               pair['longitude'],
+                               client.get(entity_id),
+                               pair['precision'])
 
     def commonsMedia__string(self,
                              client: Client,